+++ src/main/java/com/team12_week5/App.java
@@ -1,4 +1,4 @@
-package com.team12_week5.BillProgram;
+package com.team12_week5;
 
 /**
  * Hello world!
