+++ src/main/java/com/team12_week5/BasicPlan.java
@@ -1,4 +1,4 @@
-package com.team12_week5.BillProgram;
+package com.team12_week5;
 
 public class BasicPlan {
 	protected double basicFee=0;
