+++ src/test/java/com/team12_week5/AppTest.java
@@ -1,4 +1,4 @@
-package com.team12_week5.BillProgram;
+package com.team12_week5;
 
 import static org.junit.Assert.*;
